Refactor group routes into their own file.
