+++ app.js
@@ -10,6 +10,9 @@ mongoose.connect("mongodb://localhost/AddressBook", {useNewUrlParser: true});
 let Person = require("./models/Person");
 let Group = require("./models/Group");
 
+let GroupRoutes = require("./routes/groups");
+app.use(GroupRoutes);
+
 // Use EJS for templating
 app.set("view engine", "ejs");
 
@@ -141,77 +144,6 @@ app.post("/people", (req, res) => {
 	});
 });
 
-app.get("/group/new", (req, res) => {
-	res.render("group/new");
-});
-
-app.post("/group", (req, res) => {
-
-	let colour = hexToRgb(req.body.colour);
-
-	// From: https://stackoverflow.com/questions/1855884/determine-font-color-based-on-background-color
-	let luminance = ( 0.299 * colour.r + 0.587 * colour.g + 0.114 * colour.b)/255;
-
-	let fontColour = (luminance > 0.5) ? "#000000" : "#FFFFFF";
-
-	Group.create({
-		name: req.body.name,
-		colour:req.body.colour,
-		fontColour
-	}, (err) => {
-		if(err) return res.redirect("/error");
-		res.redirect("/");
-	})
-});
-
-app.get("/group/:id/edit", (req, res) => {
-	Group.findById(req.params.id)
-	.then(group => {
-		res.render("group/edit", {group});
-	})
-	.catch(err => {
-		res.redirect("/error");
-	});
-});
-
-app.put("/group/:id", (req, res) => {
-	let colour = hexToRgb(req.body.colour);
-
-	let luminance = ( 0.299 * colour.r + 0.587 * colour.g + 0.114 * colour.b)/255;
-
-	let fontColour = (luminance > 0.5) ? "#000000" : "#FFFFFF";
-
-	Group.updateOne({_id: req.params.id}, {
-		name: req.body.name,
-		colour:req.body.colour,
-		fontColour
-	})
-	.then(group => {
-		res.redirect("/");
-	}).catch(err => {
-		res.redirect("/error");
-	});
-});
-
-app.delete("/group/:id", (req, res) => {
-	let personPromise;
-
-	if(req.query.deleteMembers){
-		// Delete all members with this group
-		personPromise = Person.deleteMany({groups: mongoose.Types.ObjectId(req.params.id)});
-	} else{
-		// Remove group reference from people
-		personPromise = Person.updateMany({groups: req.params.id}, {$pull: {groups: req.params.id}});
-	}
-
-	// Delete the group
-	Promise.all([Group.deleteOne({_id:req.params.id}), personPromise])
-	.then(data => {
-		res.redirect("/");
-	}).catch(err => {
-		res.redirect("/error");
-	})
-});
 
 app.get("/error", (req, res) => {
 	res.render("error");
